Move Arbitrary for Constraint to random/constraint.rs
