+++ rust/ommx/src/convert/instance.rs
@@ -1,5 +1,5 @@
 use crate::{
-    random::{arbitrary_decision_variables, FunctionParameters},
+    random::{arbitrary_constraints, arbitrary_decision_variables, FunctionParameters},
     sorted_ids::{BinaryIdPair, BinaryIds},
     v1::{
         decision_variable::Kind,
@@ -16,8 +16,6 @@ use std::{
     borrow::Cow,
     collections::{BTreeMap, BTreeSet, HashMap, HashSet},
 };
-
-use super::constraint::arbitrary_constraints;
 
 impl Instance {
     pub fn objective(&self) -> Cow<Function> {
